+++ src/internal/m365/onedrive/restore_test.go
@@ -329,47 +329,57 @@ func (suite *RestoreUnitSuite) TestAugmentRestorePaths_DifferentRestorePath() {
 func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 	const mndiID = "mndi-id"
 
+	type counts struct {
+		skip    int64
+		replace int64
+		new     int64
+	}
+
 	table := []struct {
 		name          string
-		collisionKeys map[string]api.DriveCollisionItem
+		collisionKeys map[string]api.DriveItemIDType
 		onCollision   control.CollisionPolicy
 		deleteErr     error
 		expectSkipped assert.BoolAssertionFunc
 		expectMock    func(*testing.T, *mock.RestoreHandler)
+		expectCounts  counts
 	}{
 		{
 			name:          "no collision, copy",
-			collisionKeys: map[string]api.DriveCollisionItem{},
+			collisionKeys: map[string]api.DriveItemIDType{},
 			onCollision:   control.Copy,
 			expectSkipped: assert.False,
 			expectMock: func(t *testing.T, rh *mock.RestoreHandler) {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 0, 1},
 		},
 		{
 			name:          "no collision, replace",
-			collisionKeys: map[string]api.DriveCollisionItem{},
+			collisionKeys: map[string]api.DriveItemIDType{},
 			onCollision:   control.Replace,
 			expectSkipped: assert.False,
 			expectMock: func(t *testing.T, rh *mock.RestoreHandler) {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 0, 1},
 		},
 		{
 			name:          "no collision, skip",
-			collisionKeys: map[string]api.DriveCollisionItem{},
+			collisionKeys: map[string]api.DriveItemIDType{},
 			onCollision:   control.Skip,
 			expectSkipped: assert.False,
 			expectMock: func(t *testing.T, rh *mock.RestoreHandler) {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 0, 1},
 		},
 		{
 			name: "collision, copy",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {ItemID: mndiID},
 			},
 			onCollision:   control.Copy,
@@ -378,10 +388,11 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 0, 1},
 		},
 		{
 			name: "collision, replace",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {ItemID: mndiID},
 			},
 			onCollision:   control.Replace,
@@ -391,10 +402,11 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.True(t, rh.CalledDeleteItem, "new item deleted")
 				assert.Equal(t, mndiID, rh.CalledDeleteItemOn, "deleted the correct item")
 			},
+			expectCounts: counts{0, 1, 0},
 		},
 		{
 			name: "collision, replace - err already deleted",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {ItemID: "smarf"},
 			},
 			onCollision:   control.Replace,
@@ -404,10 +416,11 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.True(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 1, 0},
 		},
 		{
 			name: "collision, skip",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {ItemID: mndiID},
 			},
 			onCollision:   control.Skip,
@@ -416,10 +429,11 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.False(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{1, 0, 0},
 		},
 		{
 			name: "file-folder collision, copy",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {
 					ItemID:   mndiID,
 					IsFolder: true,
@@ -431,10 +445,11 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 0, 1},
 		},
 		{
 			name: "file-folder collision, replace",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {
 					ItemID:   mndiID,
 					IsFolder: true,
@@ -446,10 +461,11 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.True(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{0, 0, 1},
 		},
 		{
 			name: "file-folder collision, skip",
-			collisionKeys: map[string]api.DriveCollisionItem{
+			collisionKeys: map[string]api.DriveItemIDType{
 				mock.DriveItemFileName: {
 					ItemID:   mndiID,
 					IsFolder: true,
@@ -461,6 +477,7 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 				assert.False(t, rh.CalledPostItem, "new item posted")
 				assert.False(t, rh.CalledDeleteItem, "new item deleted")
 			},
+			expectCounts: counts{1, 0, 0},
 		},
 	}
 	for _, test := range table {
@@ -491,6 +508,8 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 			dp, err := path.ToDrivePath(dpp)
 			require.NoError(t, err)
 
+			ctr := count.New()
+
 			_, skip, err := restoreItem(
 				ctx,
 				rh,
@@ -511,10 +530,14 @@ func (suite *RestoreUnitSuite) TestRestoreItem_collisionHandling() {
 					Reader: mock.FileRespReadCloser(mock.DriveFilePayloadData),
 				},
 				nil,
-				count.New())
+				ctr)
+
 			require.NoError(t, err, clues.ToCore(err))
 			test.expectSkipped(t, skip)
 			test.expectMock(t, rh)
+			assert.Equal(t, test.expectCounts.skip, ctr.Get(count.CollisionSkip), "skips")
+			assert.Equal(t, test.expectCounts.replace, ctr.Get(count.CollisionReplace), "replaces")
+			assert.Equal(t, test.expectCounts.new, ctr.Get(count.NewItemCreated), "new items")
 		})
 	}
 }