+++ src/pkg/fault/fault.go
@@ -137,9 +137,9 @@ func (e *Bus) logAndAddRecoverable(ctx context.Context, err error, skip int) {
 	isFail := e.addRecoverableErr(err)
 
 	if isFail {
-		log.Error("recoverable error")
+		log.Errorf("recoverable error: %v", err)
 	} else {
-		log.Info("recoverable error")
+		log.Infof("recoverable error: %v", err)
 	}
 }